reorganize plotting features
