+++ screenpro/dashboard/__init__.py
@@ -1,3 +1,9 @@
+## Copyright (c) 2022-2024 ScreenPro2 Development Team.
+## All rights reserved.
+## Gilbart Lab, UCSF / Arc Institute.
+## Multi-Omics Tech Center, Arc Insititue.
+
+
 import numpy as np
 import pandas as pd
 import bokeh
